+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.9.3] - 2025-07-27
+
+- Added support for parsing tender documents
+- Added support for parsing contract documents
+- Added support for parsing organization documents
+- Added support for parsing generic documents
+- Added support for parsing attachments
+- Added support for parsing tender documents
+- Added support for parsing contract documents
+- Added support for parsing organization documents
+
 ## [0.9.0] - 2025-07-15
 
 - Initial release